Move dataconnect tests (#7318)
